일부 버그 수정
아이템 세부창에서 경험치 표시되지않는 것 수정
몬스터 풀링 시 애니메이션 초기화 추가

+++ PortfolioE/Source/PortfolioE/Private/Widget/POEItemDetailViewWidget.cpp
@@ -37,7 +37,7 @@ void UPOEItemDetailViewWidget::InitView()
 			SetAttackValueText((int)EquipItemData->GetAttackValue());
 			SetHpValueText((int)EquipItemData->GetHpValue());
 			SetSpeedValueText((int)EquipItemData->GetMoveSpeedValue());
-			ExpBar->SetPercent(EquipItemData->GetCurrentExp() / EquipItemData->GetRequireExp());
+			ExpBar->SetPercent((float)EquipItemData->GetCurrentExp() / EquipItemData->GetRequireExp());
 			ItemTypeText->SetText(FText::FromString(EquipItemData->IsPassive ? TEXT("(PASSIVE)") : TEXT("(ACTIVE)")));
 			ItemLevelText->SetText(FText::FromString(FString::Printf(TEXT("Level: %d"), ItemData->GetItemLevel())));
 		}
